pyproject.toml
[project]
name = "agentforce-mcp-server"
version = "0.1.0"
description = "MCP Server that communicates to the Agentforce API, thereby turning an agent into an MCP tool"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.12.4",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
]
